Correct CORS for HEAD requests
HEAD requests are automatically processed by Spring as GET requests with response body stripped. But when CORS headers (Sec-Fetch-*) are passed to a HEAD request, the redirection is not taken into account and error 403 is returned. To fix this, we need to explicitly add HEAD method to the mapping of all these pages.