Introduce a new formula for distance to a contact
Instead of closeness to a contact in the range [0..+inf) use the distance
calculated as follows.

  distance = constant_part - variable_part

The constant part is:

  0.75, for friends;
  1, for subscriptions;
  3, for other nodes;
  4, for blocked nodes.

The variable part is tanh(sum_of_favors / 100), where sum_of_favors is
a sum of points got for every "favor" the node did to the contact: like,
comment etc. The favor "decays" over time according to the formula:

  a(t) = A * (1 - t^2 / T^2),

where A is the initial value of the favor, t - time passed and T -
the period of the total decay. After T is passed, the favor is no more
relevant.

See FavorType enum for the list of favors and their values and decay periods.